Changed datatypes of TorrentFile to use uint8_t instead of int. Updated Hash function and properly imported crypto++ lib

+++ torrentfile/TorrentFile.cpp
@@ -5,8 +5,8 @@
 #include "TorrentFile.h"
 
 TorrentFile::TorrentFile(const std::string& announce,
-                         const std::array<int, 20>& infoHash,
-                         const std::vector<std::array<int, 20>>& pieceHashes,
+                         const std::array<uint8_t, 20>& infoHash,
+                         const std::vector<std::array<uint8_t, 20>>& pieceHashes,
                          int pieceLength, int length,
                          const std::string& name)
         : announce(announce),
